Create AddItem form
Put footer on wardrobe table view controller that
allows you to add a new item to your wardrobe.
Create segues for getting to and from this form.

Add root view controller that has selections for the
different parts of the app. Create Core Data model, but
haven't implemented it yet.