Complete Core Data implementation
Finish DataManager class.  Add a delegate that
will follow the DataManagerDelegate Protocol and
handle creation of the data if it doesn't exist yet, and
it will provide an entity name and a model name for
the data Manager. Move Core Data functions into this
class.  Add fetching of data method.

Add addItem method to WardrobeItem that creates a
WardrobeItemMO object using the given Data Manager.
Change model initialization to initialize data from the data
manager. Change adding of items to create the Model Object
and save the context.