Add Item Controller added to, small changes
Change the choosing of where the Item was purchased
to simply use a text field.  Add color text fields array to
eventually be able to have the user add color text fields
and put in a bunch of colors.  Made changing of objects'
interactability for the detail view to be a loop.

Small color changes made to make app's colors
more dynamic and easy to change.