Made Trends graph represent actual data
Trends View Controller now gets data for amount of
clothes purchased in specified time period and puts it
on the displayed graph. Also messing around a bit with
graph settings.

Create clothesBy function in model that gets a dictionary
mapping a string to an int, representing the amount of clothes
in the given time period, either year month, or week.  Still need
to implement week. Also make favoriteStore computed variable
that represents the favorite store of the user.