Add food to game
Change painting function to draw the food also.
In move function, check if snake is at the food,
then have the snake eat the food to get longer,
and randomly generate a new food.

Change initialSnakePosition variable to be
initialized dynamically with a list comprehension
based on the game width and height.